backend_server/requirements.txt
django
djangorestframework
django-cors-headers
PyPDF2
openai
python-dotenv
